requirements.txt
fastapi
uvicorn
pydantic
httpx
aiosmtplib
langchain
python-dotenv
